Rehydration Rate Calculation
By defult, the rehydration rate is calcualted using the 25 seconds of data caputred before or after sliding.  This number is optimal based on a previous convergence study of the rehydration rate calculation.  The user can override this by adding name value pairs to input a custom amount of time  or the ycan choose to input exactly how many data points they want to fit.