Bug fixes and added features
tribmetasub - Most new changes are here.  Changed the method of finding a match to using strfind to match the pattern.  Erronesouly detected segemnts are filtered out first, and a selection is made between cases of multiple matches with an additional logic statement.

tribsegprocess - outputs the the segmented tribology data cell array variable in aaddition to the cleaned version of the entire dataset.

tribplotyy - suppressed command line output

tribsegcombine - sometimes a NaN occured at the last data point of a segment.  This was used in calcualted later adjusted time points whic hcaused them all to be NaNs.  Changed to use the last time point recorded before the NaN value to prevent this error.